Donor Trace App

+++ DonorTraceAPI/Dto/DonorDto.cs
@@ -20,8 +20,8 @@ namespace DonorTraceAPI.Dto
       
         public Donor.Sex Gender { get; set; }
         public string Location { get; set; }
-        public int? Longitude { get; set; }
-        public int? Latitude { get; set; }
+        public double? Longitude { get; set; }
+        public double? Latitude { get; set; }
         public int RegionId { get; set; }
     
         public string Phone { get; set; }
