+++ DonorTraceAPI/Models/Donor.cs
@@ -18,12 +18,11 @@ namespace DonorTraceAPI.Models
         public string FirstName { get; set; }
         [Required]
         public string LastName { get; set; }
-
         [Required]
         public Sex Gender { get; set; }
         public string Location { get; set; }
-        public int? Longitude { get; set; }
-        public int? Latitude { get; set; }
+        public double? Longitude { get; set; }
+        public double? Latitude { get; set; }
         public int RegionId { get; set; }
         [Required, MaxLength(10)]
         public string Phone { get; set; }
