+++ DonorTraceMobile/DonorTraceMobile/Models/Donor.cs
@@ -19,8 +19,8 @@ namespace DonorTraceMobile.Models
 
         public string Gender { get; set; }
         public string Location { get; set; }
-        public int? Longitude { get; set; }
-        public int? Latitude { get; set; }
+        public double? Longitude { get; set; }
+        public double? Latitude { get; set; }
         public int RegionId { get; set; }
 
         public string Phone { get; set; }
